requirements.txt
Flask>=2.3.2
flask-cors>=4.0.0
pydantic>=2.1.1
pdf2image>=1.16.3
PyJWT>=2.8.0
pymongo>=4.4.1
python-dotenv>=1.0.0
